+++ x/staking/types/msg_test.go
@@ -6,6 +6,11 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/tendermint/tendermint/crypto"
 
+	"github.com/cosmos/cosmos-sdk/codec"
+	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
+	cryptocodec "github.com/cosmos/cosmos-sdk/crypto/codec"
+	"github.com/cosmos/cosmos-sdk/crypto/keys/ed25519"
+	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
@@ -13,6 +18,38 @@ var (
 	coinPos  = sdk.NewInt64Coin(sdk.DefaultBondDenom, 1000)
 	coinZero = sdk.NewInt64Coin(sdk.DefaultBondDenom, 0)
 )
+
+func TestMsgDecode(t *testing.T) {
+	registry := codectypes.NewInterfaceRegistry()
+	cryptocodec.RegisterInterfaces(registry)
+	RegisterInterfaces(registry)
+	cdc := codec.NewProtoCodec(registry)
+
+	// firstly we start testing the pubkey serialization
+
+	pk1bz, err := codec.MarshalAny(cdc, pk1)
+	require.NoError(t, err)
+	var pkUnmarshaled cryptotypes.PubKey
+	err = codec.UnmarshalAny(cdc, &pkUnmarshaled, pk1bz)
+	require.NoError(t, err)
+	require.True(t, pk1.Equals(pkUnmarshaled.(*ed25519.PubKey)))
+
+	// now let's try to serialize the whole message
+
+	commission1 := NewCommissionRates(sdk.ZeroDec(), sdk.ZeroDec(), sdk.ZeroDec())
+	msg, err := NewMsgCreateValidator(valAddr1, pk1, coinPos, Description{}, commission1, sdk.OneInt())
+	require.NoError(t, err)
+	msgSerialized, err := codec.MarshalAny(cdc, msg)
+	require.NoError(t, err)
+
+	var msgUnmarshaled sdk.Msg
+	err = codec.UnmarshalAny(cdc, &msgUnmarshaled, msgSerialized)
+	require.NoError(t, err)
+	msg2, ok := msgUnmarshaled.(*MsgCreateValidator)
+	require.True(t, ok)
+	require.True(t, msg.Value.IsEqual(msg2.Value))
+	require.True(t, msg.Pubkey.Equal(msg2.Pubkey))
+}
 
 // test ValidateBasic for MsgCreateValidator
 func TestMsgCreateValidator(t *testing.T) {
@@ -41,8 +78,10 @@ func TestMsgCreateValidator(t *testing.T) {
 	}
 
 	for _, tc := range tests {
+		t.Logf("Test: %s, pk=%t", tc.name, tc.pubkey)
 		description := NewDescription(tc.moniker, tc.identity, tc.website, tc.securityContact, tc.details)
-		msg := NewMsgCreateValidator(tc.validatorAddr, tc.pubkey, tc.bond, description, tc.CommissionRates, tc.minSelfDelegation)
+		msg, err := NewMsgCreateValidator(tc.validatorAddr, tc.pubkey, tc.bond, description, tc.CommissionRates, tc.minSelfDelegation)
+		require.NoError(t, err)
 		if tc.expectPass {
 			require.Nil(t, msg.ValidateBasic(), "test: %v", tc.name)
 		} else {