+++ x/staking/types/validator.go
@@ -14,6 +14,7 @@ import (
 	"gopkg.in/yaml.v2"
 
 	"github.com/cosmos/cosmos-sdk/codec"
+	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
@@ -39,15 +40,15 @@ var _ ValidatorI = Validator{}
 
 // NewValidator constructs a new Validator
 //nolint:interfacer
-func NewValidator(operator sdk.ValAddress, pubKey crypto.PubKey, description Description) Validator {
-	var pkStr string
-	if pubKey != nil {
-		pkStr = sdk.MustBech32ifyPubKey(sdk.Bech32PubKeyTypeConsPub, pubKey)
+func NewValidator(operator sdk.ValAddress, pubKey crypto.PubKey, description Description) (Validator, error) {
+	pkAny, err := codectypes.PackAny(pubKey)
+	if err != nil {
+		return Validator{}, err
 	}
 
 	return Validator{
 		OperatorAddress:   operator.String(),
-		ConsensusPubkey:   pkStr,
+		ConsensusPubkey:   pkAny,
 		Jailed:            false,
 		Status:            Unbonded,
 		Tokens:            sdk.ZeroInt(),
@@ -57,7 +58,7 @@ func NewValidator(operator sdk.ValAddress, pubKey crypto.PubKey, description Des
 		UnbondingTime:     time.Unix(0, 0).UTC(),
 		Commission:        NewCommission(sdk.ZeroDec(), sdk.ZeroDec(), sdk.ZeroDec()),
 		MinSelfDelegation: sdk.OneInt(),
-	}
+	}, nil
 }
 
 // String implements the Stringer interface for a Validator object.
@@ -87,13 +88,17 @@ func (v Validators) ToSDKValidators() (validators []ValidatorI) {
 }
 
 // ToTmValidators casts all validators to the corresponding tendermint type.
-func (v Validators) ToTmValidators() []*tmtypes.Validator {
+func (v Validators) ToTmValidators() ([]*tmtypes.Validator, error) {
 	validators := make([]*tmtypes.Validator, len(v))
+	var err error
 	for i, val := range v {
-		validators[i] = val.ToTmValidator()
+		validators[i], err = val.ToTmValidator()
+		if err != nil {
+			return nil, err
+		}
 	}
 
-	return validators
+	return validators, nil
 }
 
 // Sort Validators sorts validator array in ascending operator address order
@@ -118,9 +123,19 @@ func (v Validators) Swap(i, j int) {
 	v[j] = it
 }
 
+// UnpackInterfaces implements UnpackInterfacesMessage.UnpackInterfaces
+func (v Validators) UnpackInterfaces(c codectypes.AnyUnpacker) error {
+	for i := range v {
+		if err := v[i].UnpackInterfaces(c); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
 // return the redelegation
-func MustMarshalValidator(cdc codec.BinaryMarshaler, validator Validator) []byte {
-	return cdc.MustMarshalBinaryBare(&validator)
+func MustMarshalValidator(cdc codec.BinaryMarshaler, validator *Validator) []byte {
+	return cdc.MustMarshalBinaryBare(validator)
 }
 
 // unmarshal a redelegation from a store value
@@ -233,7 +248,11 @@ func (d Description) EnsureLength() (Description, error) {
 // ABCIValidatorUpdate returns an abci.ValidatorUpdate from a staking validator type
 // with the full validator power
 func (v Validator) ABCIValidatorUpdate() abci.ValidatorUpdate {
-	pk, err := encoding.PubKeyToProto(v.GetConsPubKey())
+	consPk, err := v.TmConsPubKey()
+	if err != nil {
+		panic(err)
+	}
+	pk, err := encoding.PubKeyToProto(consPk)
 	if err != nil {
 		panic(err)
 	}
@@ -247,7 +266,11 @@ func (v Validator) ABCIValidatorUpdate() abci.ValidatorUpdate {
 // ABCIValidatorUpdateZero returns an abci.ValidatorUpdate from a staking validator type
 // with zero power used for validator updates.
 func (v Validator) ABCIValidatorUpdateZero() abci.ValidatorUpdate {
-	pk, err := encoding.PubKeyToProto(v.GetConsPubKey())
+	consPk, err := v.TmConsPubKey()
+	if err != nil {
+		panic(err)
+	}
+	pk, err := encoding.PubKeyToProto(consPk)
 	if err != nil {
 		panic(err)
 	}
@@ -259,8 +282,12 @@ func (v Validator) ABCIValidatorUpdateZero() abci.ValidatorUpdate {
 }
 
 // ToTmValidator casts an SDK validator to a tendermint type Validator.
-func (v Validator) ToTmValidator() *tmtypes.Validator {
-	return tmtypes.NewValidator(v.GetConsPubKey(), v.ConsensusPower())
+func (v Validator) ToTmValidator() (*tmtypes.Validator, error) {
+	consPk, err := v.TmConsPubKey()
+	if err != nil {
+		return nil, err
+	}
+	return tmtypes.NewValidator(consPk, v.ConsensusPower()), nil
 }
 
 // SetInitialCommission attempts to set a validator's initial commission. An
@@ -415,14 +442,24 @@ func (v Validator) RemoveDelShares(delShares sdk.Dec) (Validator, sdk.Int) {
 
 // MinEqual defines a more minimum set of equality conditions when comparing two
 // validators.
-func (v Validator) MinEqual(other Validator) bool {
-	return v.ConsensusPubkey == other.ConsensusPubkey &&
-		(v.OperatorAddress == other.OperatorAddress) &&
+func (v *Validator) MinEqual(other *Validator) bool {
+	return v.OperatorAddress == other.OperatorAddress &&
 		v.Status == other.Status &&
 		v.Tokens.Equal(other.Tokens) &&
 		v.DelegatorShares.Equal(other.DelegatorShares) &&
-		v.Description == other.Description &&
-		v.Commission.Equal(other.Commission)
+		v.Description.Equal(other.Description) &&
+		v.Commission.Equal(other.Commission) &&
+		v.Jailed == other.Jailed &&
+		v.MinSelfDelegation.Equal(other.MinSelfDelegation) &&
+		v.ConsensusPubkey.Equal(other.ConsensusPubkey)
+
+}
+
+// Equal checks if the receiver equals the parameter
+func (v *Validator) Equal(v2 *Validator) bool {
+	return v.MinEqual(v2) &&
+		v.UnbondingHeight == v2.UnbondingHeight &&
+		v.UnbondingTime.Equal(v2.UnbondingTime)
 }
 
 func (v Validator) IsJailed() bool        { return v.Jailed }
@@ -438,24 +475,43 @@ func (v Validator) GetOperator() sdk.ValAddress {
 	}
 	return addr
 }
-func (v Validator) GetConsPubKey() crypto.PubKey {
+
+// TmConsPubKey casts Validator.ConsensusPubkey to crypto.PubKey
+func (v Validator) TmConsPubKey() (crypto.PubKey, error) {
+	pk, ok := v.ConsensusPubkey.GetCachedValue().(cryptotypes.PubKey)
+	if !ok {
+		return nil, sdkerrors.Wrapf(sdkerrors.ErrInvalidType, "Expecting crypto.PubKey, got %T", pk)
+	}
+
 	// The way things are refactored now, v.ConsensusPubkey is sometimes a TM
 	// ed25519 pubkey, sometimes our own ed25519 pubkey. This is very ugly and
 	// inconsistent.
 	// Luckily, here we coerce it into a TM ed25519 pubkey always, as this
 	// pubkey will be passed into TM (eg calling encoding.PubKeyToProto).
-	pk := sdk.MustGetPubKeyFromBech32(sdk.Bech32PubKeyTypeConsPub, v.ConsensusPubkey)
-
 	if intoTmPk, ok := pk.(cryptotypes.IntoTmPubKey); ok {
-		return intoTmPk.AsTmPubKey()
+		return intoTmPk.AsTmPubKey(), nil
 	}
-
-	return pk
+	return nil, sdkerrors.Wrapf(sdkerrors.ErrInvalidPubKey, "Logic error: ConsensusPubkey must be an SDK key and SDK PubKey types must be convertible to tendermint PubKey; got: %T", pk)
 }
-func (v Validator) GetConsAddr() sdk.ConsAddress  { return sdk.ConsAddress(v.GetConsPubKey().Address()) }
+
+// GetConsAddr extracts Consensus key address
+func (v Validator) GetConsAddr() (sdk.ConsAddress, error) {
+	pk, err := v.TmConsPubKey()
+	if err != nil {
+		return sdk.ConsAddress{}, err
+	}
+	return sdk.ConsAddress(pk.Address()), nil
+}
+
 func (v Validator) GetTokens() sdk.Int            { return v.Tokens }
 func (v Validator) GetBondedTokens() sdk.Int      { return v.BondedTokens() }
 func (v Validator) GetConsensusPower() int64      { return v.ConsensusPower() }
 func (v Validator) GetCommission() sdk.Dec        { return v.Commission.Rate }
 func (v Validator) GetMinSelfDelegation() sdk.Int { return v.MinSelfDelegation }
 func (v Validator) GetDelegatorShares() sdk.Dec   { return v.DelegatorShares }
+
+// UnpackInterfaces implements UnpackInterfacesMessage.UnpackInterfaces
+func (v Validator) UnpackInterfaces(unpacker codectypes.AnyUnpacker) error {
+	var pk crypto.PubKey
+	return unpacker.UnpackAny(v.ConsensusPubkey, &pk)
+}